+++ frontend/src/hooks/api/profile/useProfileEdit.ts
@@ -1,6 +1,9 @@
-import { useEffect } from 'react';
 import { User } from '@prisma/client';
-import { UseMutateAsyncFunction } from '@tanstack/react-query';
+import {
+  UseMutateAsyncFunction,
+  UseMutationResult,
+  useQueryClient,
+} from '@tanstack/react-query';
 import { isAxiosError } from 'axios';
 import { useCustomToast } from 'hooks/utils/useCustomToast';
 import { usePostApi } from '../generics/usePostApi';
@@ -21,28 +24,29 @@ export type EditProfile = UseMutateAsyncFunction<
   unknown
 >;
 
-export const useProfileEdit = (): {
-  editProfile: EditProfile;
-  isLoading: boolean;
-  isSuccess: boolean;
-} => {
-  const {
-    mutateAsync: editProfile,
-    isLoading,
-    isSuccess,
-    isError,
-    error,
-  } = usePostApi<ProfileFormData, ProfileEditResBody>(`/users/me/profile`, [
-    ['/users/me/profile'],
-    ['/game/matches'],
-  ]);
-
+export const useProfileEdit = (): Omit<
+  UseMutationResult<ProfileEditResBody, unknown, ProfileFormData, unknown>,
+  'mutateAsync'
+> & { editProfile: EditProfile } => {
+  const queryClient = useQueryClient();
   const { customToast } = useCustomToast();
-  useEffect(() => {
-    if (isError && isAxiosError<{ message: string }>(error)) {
-      customToast({ description: error.response?.data.message });
-    }
-  }, [isError, error, customToast]);
 
-  return { editProfile, isLoading, isSuccess };
+  const { mutateAsync: editProfile, ...useMutationResult } = usePostApi<
+    ProfileFormData,
+    ProfileEditResBody
+  >(`/users/me/profile`, {
+    onSuccess: () => {
+      const queryKeys = [['/users/me/profile'], ['/game/matches']];
+      queryKeys.forEach((queryKey) => {
+        void queryClient.invalidateQueries({ queryKey });
+      });
+    },
+    onError: (error) => {
+      if (isAxiosError<{ message: string }>(error)) {
+        customToast({ description: error.response?.data.message });
+      }
+    },
+  });
+
+  return { editProfile, ...useMutationResult };
 };
